Allocate tx_queue for packet lcore
Packet lcore needs a tx_queue because agent would like to transmit
packets out on any interface it desires, and all of them will arrive
in the packet lcore. This tx queue will be a hardware queue, if the
number of available hardware queues are more than forwarding lcores.
Else, it will be a software ring.

Also, the netlink header flag should be set to zero for the last message
of a bunch. We were setting it to NLMSG_DONE, which is actually a type
of netlink message rather than a flag. While, as per netlink man page
we are supposed to send an empty message with type NLMSG_DONE as the
last message of the bunch, for now we will just set the flags to zero
and move on.