Revert a change that was deleted by the previous commit.
The number of usable hardware tx queues is a minimum of actual available
tx queues and the number of forwarding lcores. The forwarding lcore number
doesn't contain packet lcore and hence if the number of available queues
are more than the forwarding lcores, we will not use a hardware queue for
packets from agent. Add 1 to account for packet lcore while calculating
the usable queues.

Also make the netlink and the packet loop honor the stop flag, so that
when an interrupt signal is sent, the application exits