This commit revokes commit: ec5a8da262c6cf24a3fc5f9357c0766fa84473cd
Above commit has been fixing warning that have been appearing when
starting vrouter with 1GBit cards; unfrotunately it causes error while
setting up the 10Gbit for operation (it show up in vrouter startup log).

This shall be further investigated as, probably, both the error and
the warning are not wanted.