Fix log information while adding KNI device.
While adding vhost interface in xconnect, there is a line in log saying:
"Adding vif # KNI device # at eth device # MAC #:#:#:#:#:#". If the interface
was addedd using a --pci flag and --xconnect was pointing to a PCI address,
the MAC address in log was displayed as 00:00:00:00:00:00 and incorrect
device number was displayed. The fix solves this issue by looking for port_id
directly in the vif struct or by using the dpdk_find_port_id_by_pci_addr(),
regarding the VIF_FLAG_PMD flag.