Compute the number of cores to run vRouter on.
Implement dpdk_compute_core_mask() and dpdk_argv_update() functions,
to compute an affinity mask and pass it to the dpdk_argv[] array
containing parameters to EAL initialization.

This lets us to start vRouter with a custom affinity mask set by taskset,
in opposition to the previous solution, where the default mask 0x3f was
hardcoded and used to initialize EAL with.