Another implementation of ring allocation/deallocation.
Memory for a ring is allocated manually using rte_malloc_socket()
and initialized with rte_ring_init(). This approach lets to deallocate
memory with rte_free(), what was impossible while creating rings with
rte_ring_create(), where memory was allocated with rte_memzone_reserve().

Rings are dealocated with dpdk_ring_free(), so there is no need to implement
a mechanism of re-using the memory. It's just allocated once again if needed.

dpdk_ring_allocate() has been rewritten to be a generic function.